Implemented auto selection of a saved adapter
The chosen adapter can not be set to be remembered the next time the
program is started. The program will be started with this adapter if
it's available, or display the normal interface if the adapter could not
be found. The feature can be disabled or re-enabled from settings.
(Implements feature request #6)